requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
idna==3.2
multidict==5.1.0
python-dotenv==0.19.0
pytz==2021.1
typing-extensions==3.10.0.2
yarl==1.6.3
